chore: drop net3.1 & net5.0 from LinuxRunSamples. add net7.0
