--- samples/LinuxRunSamples/5.0/BuildTest.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="SkiaSharp.QrCode" Version="0.4.1" />
-    <!-- require libfontconfig1 https://github.com/mono/SkiaSharp/issues/964#issuecomment-549385484 -->
-    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.80.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="samples\*.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-</Project>
