+++ samples/LinuxRunSamples/7.0/entrypoint.sh
